requirements.txt
llama-index
llama-cloud
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-clip
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-readers-structured-data
llama-parse
llama-index-readers-microsoft-onedrive
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-vector-stores-qdrant
llama-index-readers-structured-data
aiohappyeyeballs
sentence-transformers
aiohttp
aioredis
aiosignal
alembic
annotated-types
anyio
openpyxl
torch
python-pptx
Pillow
xlrd
docx2txt
asgiref
asttokens
async-timeout
attrs
backoff
python-multipart
bcrypt
beautifulsoup4
bidict
build
cachetools
certifi
cffi
charset-normalizer
click
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
coloredlogs
cryptography
dataclasses-json
decorator
Deprecated
dirtyjson
distro
dnspython
durationpy
ecdsa
email_validator
executing
fastapi
fastapi-limiter
filelock
filetype
flatbuffers
frozenlist
fsspec
ftfy
google-auth
googleapis-common-protos
greenlet
grpcio
h11
httpcore
httptools
httpx
humanfriendly
idna
importlib_metadata
importlib_resources
ipython
jedi
Jinja2
jiter
joblib
jsonpatch
jsonpointer
kubernetes
Mako
markdown-it-py
MarkupSafe
marshmallow
matplotlib-inline
mdurl
mmh3
monotonic
mpmath
multidict
mypy-extensions
nest-asyncio
networkx
nltk
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pandas
parso
passlib
pdfminer.six
pdfplumber
pexpect
pillow
pinecone-client
pinecone-plugin-inference
pinecone-plugin-interface
posthog
prompt_toolkit
propcache
protobuf
psutil
psycopg2-binary
ptyprocess
pure_eval
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
PyJWT
pyotp
pypdf
pypdfium2
PyPika
pyproject_hooks
python-dateutil
python-dotenv
python-engineio
python-jose
python-multipart
python-socketio
pytz
PyYAML
qrcode
redis
regex
requests
requests-oauthlib
requests-toolbelt
rich
rsa
setuptools
shellingham
simple-websocket
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
striprtf
sympy
tenacity
tiktoken
tokenizers
torch
torchvision
tqdm
traitlets
triton
typer
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
uvloop
watchfiles
wcwidth
websocket-client
websockets
wrapt
wsproto
yarl
zipp
